Moved the Skia Shape View drawing into an extension
