+++ src/Comet.Skia/SkiaShapeView.cs
@@ -14,120 +14,12 @@ namespace Comet.Skia
 
         public override void Draw(SKCanvas canvas, RectangleF rect)
         {
-            canvas.Clear(SKColors.White);
-
-            if (Shape != null)
-            {
-                SKPaint strokePaint = null;
-                SKPaint fillPaint = null;
-                SKShader shader = null;
-                object fill = null;
-                float lineWidth = 0;
-                
-                var drawingStyle = Shape.GetDrawingStyle(this, DrawingStyle.StrokeFill);
-                
-                switch (drawingStyle)
-                {
-                    case DrawingStyle.StrokeFill:
-                        strokePaint = new SKPaint();
-                        strokePaint.StrokeWidth = lineWidth = Shape.GetLineWidth(this, 1);
-                        strokePaint.Color = Shape.GetStrokeColor(this, Color.Black).ToSKColor();
-                        strokePaint.Style = SKPaintStyle.Stroke;
-                        fillPaint = new SKPaint();
-                        fillPaint.Style = SKPaintStyle.Fill;
-                        fill = Shape.GetFill(this);
-                        break;
-                    case DrawingStyle.Stroke:
-                        strokePaint = new SKPaint();
-                        strokePaint.StrokeWidth = lineWidth = Shape.GetLineWidth(this, 1);
-                        strokePaint.Color = Shape.GetStrokeColor(this, Color.Black).ToSKColor();
-                        strokePaint.Style = SKPaintStyle.Stroke;
-                        fillPaint = new SKPaint();
-                        fillPaint.Style = SKPaintStyle.Fill;
-                        break;
-                    case DrawingStyle.Fill:
-                        fillPaint = new SKPaint();
-                        fillPaint.Style = SKPaintStyle.Fill;
-                        fill = Shape.GetFill(this);
-                        break;
-                }
-                
-                var shapeBounds = new RectangleF(
-                    rect.X + (lineWidth / 2),
-                    rect.Y + (lineWidth / 2),
-                    rect.Width - lineWidth,
-                    rect.Height - lineWidth);
-                
-                var path = Shape.PathForBounds(shapeBounds).ToSKPath();
-                
-                if (fill != null && fillPaint != null)
-                {
-                    if (fill is Color color)
-                    {
-                        fillPaint.Color = color.ToSKColor();
-                        canvas.DrawPath(path, fillPaint);
-                    }
-                    else if (fill is Gradient gradient)
-                    {
-                        canvas.Save();
-                        
-                        var colors = new SKColor[gradient.Stops.Length];
-                        var stops = new float[colors.Length];
-
-                        var sortedStops = gradient.GetSortedStops();
-
-                        for (var i = 0; i < sortedStops.Length; i++)
-                        {
-                            colors[i] = sortedStops[i].Color.ToSKColor();
-                            stops[i] = sortedStops[i].Offset;
-                        }
-                        
-                        if (gradient is LinearGradient linearGradient)
-                        {
-                            var x1 = rect.X + rect.Width * linearGradient.StartPoint.X;
-                            var y1 = rect.Y + rect.Height * linearGradient.StartPoint.Y;
-
-                            var x2 = rect.X + rect.Width * linearGradient.EndPoint.X;
-                            var y2 = rect.Y + rect.Height * linearGradient.EndPoint.Y;
-
-                            shader = SKShader.CreateLinearGradient(
-                                new SKPoint(x1, y1),
-                                new SKPoint(x2, y2),
-                                colors,
-                                stops,
-                                SKShaderTileMode.Clamp);
-                        }
-                        else if (gradient is RadialGradient radialGradient)
-                        {
-                            var x1 = rect.X + rect.Width * radialGradient.Center.X;
-                            var y1 = rect.Y + rect.Height * radialGradient.Center.Y;
-                            var r = radialGradient.EndRadius;
-                            
-                            shader = SKShader.CreateRadialGradient(
-                                new SKPoint(x1, y1),
-                                r,
-                                colors,
-                                stops,
-                                SKShaderTileMode.Clamp);
-                        }
-
-                        if (shader != null)
-                        {
-                            fillPaint.Shader = shader;
-                            canvas.DrawPath(path, fillPaint);
-                        }
-
-                        shader.Dispose();
-                        canvas.Restore();
-                    }
-                }
-
-                if (strokePaint != null)
-                    canvas.DrawPath(path, strokePaint);
-                
-                strokePaint?.Dispose();
-                fillPaint?.Dispose();
-            }
+            var backgroundColor = this.GetBackgroundColor(Color.White).ToSKColor();
+            canvas.Clear(backgroundColor);
+            var drawingStyle = Shape.GetDrawingStyle(this, DrawingStyle.StrokeFill);
+            var strokeColor = Shape.GetStrokeColor(this, Color.Black);
+            var strokeWidth = Shape.GetLineWidth(this, 1);
+            canvas.DrawShape(Shape,rect, drawingStyle,strokeWidth,strokeColor);
         }
     }
 }
